Restructure code into subdirectory
