+++ autobot/tests/test_bot.py
@@ -1,12 +1,17 @@
 import unittest
+
+from pathlib import Path
 from urllib.parse import urlparse, parse_qs
 
-import bot
+import autobot.bot as bot
+
 
 class TestBotMethods(unittest.TestCase):
+    def _get_files_dir(self) -> Path:
+        return Path(__file__).resolve().parent / "files"
 
     def test_generate_modmail_link(self):
-        modlink = bot.generate_modmail_link('testsr', 'some title', 'some body')
+        modlink = bot.generate_modmail_link('testsr', 'a title', 'some body')
         parsed_url = urlparse(modlink)
 
         query = parse_qs(parsed_url.query)
@@ -17,7 +22,7 @@ class TestBotMethods(unittest.TestCase):
         self.assertTrue('message' in query)
         self.assertTrue('subject' in query)
         self.assertEqual(query['message'][0], 'some body')
-        self.assertEqual(query['subject'][0], 'some title')
+        self.assertEqual(query['subject'][0], 'a title')
 
     def test_generate_empty_modmail_link(self):
         modlink = bot.generate_modmail_link('testsr')
@@ -38,12 +43,11 @@ class TestBotMethods(unittest.TestCase):
         self.assertFalse(bot.title_contains_nsfw('NsFw_title_with_spaces'))
         self.assertTrue(bot.title_contains_nsfw('nsfw leading title'))
         self.assertTrue(bot.title_contains_nsfw('Title ending with NSFW'))
-        self.assertFalse(bot.title_contains_nsfw('Title that does not contain bad words'))
+        self.assertFalse(bot.title_contains_nsfw('Title without bad words'))
         self.assertTrue(bot.title_contains_nsfw('Title with an [NSFW] tag'))
         self.assertTrue(bot.title_contains_nsfw('!NSFW!'))
         self.assertTrue(bot.title_contains_nsfw('Is this post NSFW?'))
         self.assertTrue(bot.title_contains_nsfw('Hi [Part 2] [NSFW]'))
-
 
     def test_reject_long_paragraphs(self):
         '''This test asserts that paragraphs > (length) words are rejected.'''
@@ -63,19 +67,19 @@ class TestBotMethods(unittest.TestCase):
 
     def test_live_story_paragraphs(self):
         # Soul Cancer initiated issue #13
-        with open('tests/soul_cancer.md', 'r') as sc:
+        files_dir = self._get_files_dir()
+        with open(files_dir / "soul_cancer.md", "r") as sc:
             story = sc.read()
             issues = bot.collect_formatting_issues(story)
             self.assertFalse(issues.long_paragraphs)
 
-
     def test_chezecaek_full_story(self):
         # Chezecaek initiated issue #17
-        with open('tests/chezecaek.md', 'r') as sc:
+        files_dir = self._get_files_dir()
+        with open(files_dir / "chezecaek.md", "r") as sc:
             story = sc.read()
             issues = bot.collect_formatting_issues(story)
             self.assertFalse(issues.has_codeblocks)
-
 
     def test_reject_long_paragraphs_funky_newlines(self):
         '''Test edge case long paragraphs (newlines)'''
@@ -87,7 +91,6 @@ class TestBotMethods(unittest.TestCase):
         text += ' '.join(['more'] * 100)
         issues = bot.collect_formatting_issues(text)
         self.assertFalse(issues.long_paragraphs)
-
 
     def test_contains_codeblocks(self):
         '''Test for codeblocks in a message'''
@@ -106,7 +109,6 @@ class TestBotMethods(unittest.TestCase):
 
         self.assertFalse(bot.contains_codeblocks(['']))
 
-
     def test_categorize_tags(self):
         title = 'This is a sample post (volume 1) {part 2} |part 3|'
         tags = bot.categorize_tags(title)
@@ -115,7 +117,6 @@ class TestBotMethods(unittest.TestCase):
         self.assertEqual(tags['valid_tags'][0], "volume 1")
         self.assertEqual(tags['valid_tags'][1], "part 2")
         self.assertEqual(tags['valid_tags'][2], "part 3")
-
 
     def test_update_tags(self):
         '''Test that tags like 'Update #3' and 'Update 99' are allowed'''
@@ -127,12 +128,10 @@ class TestBotMethods(unittest.TestCase):
         self.assertEqual(tags['valid_tags'][1], 'update 100')
         self.assertEqual(tags['invalid_tags'][0], 'update1')
 
-
     def test_additional_categorize_tags(self):
-        title = 'Truckers Have Some of The Best Stories: Gas Shortages and Unraveling Threads (update)'
+        title = 'Truckers Have Some of The Best Stories Threads (update)'
         tags = bot.categorize_tags(title)
         self.assertEqual(len(tags['invalid_tags']), 0)
-
 
     def test_categorize_tags_varying_case(self):
         title = 'This is a sample post (VoLuME 1) {PT 2}'
@@ -157,4 +156,3 @@ class TestBotMethods(unittest.TestCase):
         import os
         os.environ.clear()
         self.assertFalse(bool(bot.get_environment_configuration()))
-
